fix: resolve clippy lints across workspace
Fix clippy::all denials and -D warnings: collapsible ifs, map_or/is_some_and, redundant casts, manual range contains, while-let loops, nested/useless format!, missing Default impls, inspect_err, explicit clone closure. Remove unused imports and mark intentional dead code with #[allow(dead_code)].

+++ crates/protocol/src/versions/v1_7_10/play/mod.rs
@@ -1394,7 +1394,7 @@ mod packets {
                 z: src.get_i32(),
                 action_id: src.get_i8(),
                 action_param: src.get_i8(),
-                block_type: VarInt::decode(src)?.0 as i32,
+                block_type: VarInt::decode(src)?.0,
             })
         }
     }
@@ -2038,7 +2038,7 @@ mod packets {
                 x: src.get_i32(),
                 y: src.get_i8(),
                 z: src.get_i32(),
-                block_id: VarInt::decode(src)?.0 as i32,
+                block_id: VarInt::decode(src)?.0,
                 block_metadata: src.get_i8(),
             })
         }
